Mightex, Mightex_BLS: Use WinDDK 7.10 and fix build.
There are subtle include path ordering issues here. I could not get the
build to work again using the old WinDDK (2003), so switched to
WinDDK (aka WDK) 7.10 ("7600.16385.1").

Include paths must be searched in this order:
- WinDDK\...\inc\ddk (not used by Mightex)
- $(VCInstallDir)include (must override default order)
- WinDDK\...\inc\api (used by Mightex)
- $(WindowsSdkDir)\include (inherited from defaults)
(See
http://stackoverflow.com/questions/3337859/wdk-windows-driver-kit-and-vc-headers-problem
, but the approach of setting it in Visual Studio Options is of course 
fragile and should be avoided.)

The WinDDK header(s) included by Mightex cause the linker to require 
hid.lib. There are OS-specific versions of hid.lib (wxp, win7, etc.),
which is the import library for hid.dll (part of Windows, I believe).
I chose the wxp version for 32-bit, because it is presumably closest
to the old WinDDK's version. For 64-bit (not available in wxp), I
chose win7 for now; hopefully nobody is using Vista 64-bit....

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@11093 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd