Serial: Make binary logging actually readable.
The main problem was that chars >= 0x7f were being printed as chars.

Also changed to hex printing for non-graphical chars, in a format that clearly
shows that it is hex. With the old format there was no way to tell that the
bytes were printed in decimal without looking at the code here.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@11441 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd