Fix bug in property value sliders.
For integer properties with large ranges, the scroll bar is given a range of
0-1000. But when computing the slider value from the real value, the result was
being rounded down (floor) to an integer, resulting in missed steps in some
situations. The test case was an integer property with range 0-65535. The
slider would go to 65, but would not go any further by clicking on the right
arrow.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@11462 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd