Remove generated source files.
Aside from the fact that generated files generally should not be in the
repository, the contents of these files are dependent on the version of
protobuf (and version must match with the protobuf libs linked against), so can
cause confusion (esp. if they are e.g. generated on Windows and compiled on
Unix).

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@11684 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd