BaumerOptronic: Discard images captured before the acquisition.
This is a hack (the only way without a total rewrite). For some reason this
adapter continuously captures images, and grabs the most recent one when a snap
or sequence is requested. Prior to this fix, it did not wait for a new exposure
to start after SnapImage() or StartSequenceAcquisition() is called, so there
was no synchronization with the rest of Micro-Manager (esp shutters).

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@11805 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd