Correctly check for z-stack sequenceability.
The bug was that Z-stacks would fail with "Unsupported device command"
error, due to calling core.loadStageSequence() without correctly
checking core.isStageSequenceable(). This would only happen when more
than one Z slice is to be acquired, only one channel is to be acquired
(or channels are disabled), and the acquisition order is
slices-before-channels. When channels were disabled altogether, the
error would occur only if slices-before-channels was selected when
channels were previously enabled.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@12015 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd