Use DLL version of Microsoft C Runtime everywhere possible.
I.e., use /MD (and /MDd), not /MT (and /MTd).

Excluding MCL_MicroDrive and MCL_NanoDrive, which require the static version to
work with the vendor-provided libraries.

The reason for this change (other than to conserve disk space, DLL load time,
and memory footprint) is that there is a maximum of 128 for the number of
copies of the C Runtime that can be loaded at the same time [1], and that that
limit can now be reached when opening the Hardware Configuration Wizard (at
least with the 32-bit version, with several vendor DLLs installed on the
system). Without this change, the last few device adapters (in the alphabetical
list) become unavailable.

[1]: http://social.msdn.microsoft.com/Forums/windowsdesktop/en-US/3546c3c4-1b36-4552-85c5-1b3ba860ee84/how-many-dlls-can-be-loaded-in-a-process-using-loadlibrary-?forum=windowssdk

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@12190 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd