TIFF Writer: Use correct byte count, not character count.
The bug was causing corrupted TIFF files to be written when the locale was set
to e.g. zh_CN, because the matadata (and potentially other strings) could
contain non-ASCII characters whose UTF-8 representation is more than 1
byte/character. It could cause the string to overwrite neighboring data
structures (which happens to be the first IFD in the case of the summary
metadata string!).

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@12333 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd