Fix hacks involving CDeviceBase [2/8]: Marzhauser.
Eliminate the fake interface class TangoDeviceBase (from which the
concrete device classes don't actually derive, making the static_cast
from concrete device pointer to TangoDeviceBase* a dangerous hack).

Modify the TangoBase class to use MM::Core* (the call-back-to-core
interface) instead of CDeviceBase methods, so that it works without
having to have a pointer to the CDeviceBase interface of the concrete
device classes (which is what TangoDeviceBase* was used for).

Make TangoBase destructor virtual so that it actually gets called.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@13379 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd