Fix hacks involving CDeviceBase [5/8]: Piezosystem_dDrive.
Eliminate the fake interface class EVDDeviceBase (from which the
concrete device classes don't acutally derive, making the static_cast
from concrete device pointer to EVDDeviceBase* a dangerous hack).

Modify the EVDBase class to use MM::Core* (the call-back-to-core
interface) instead of CDeviceBase methods, so that it works without
having to have a pointer to the CDeviceBase interface of the concrete
device classes (which is what EVDDeviceBase* was used for).

Make EVDBase destructor virtual so that it actually gets called.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@13382 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd