Fix hacks involving CDeviceBase [4/8]: NewportCONEX.
Eliminate the fake interface class Conex_AxisDeviceBase (from which the
concrete device classes don't actually derive, making the static_cast
from concrete device pointer to Conex_AxisDeviceBase* a dangerous hack).

Modify the Conex_AxisBase class to use MM::Core* (the call-back-to-core
interface) instead of CDeviceBase methods, so that it works without
having to have a pointer to the CDeviceBase interface of the concrete
device classes (which is what Conex_AxisDeviceBase* was used for).

Make Conex_AxisBase descructor virtual so that it actually gets called.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@13381 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd