Fix hacks involving CDeviceBase [7/8]: Zaber.
Eliminate the fake interface class ZaberDeviceBase (from which the concrete
device classes don't actually derive, making the static_cast from concrete
device pointer to ZaberDeviceBase* a dangerous hack).

Modify the ZaberBase class to use MM::Core* (the call-back-to-core interface)
instead of CDeviceBase methods, so that it works without having to have a
pointer to the CDeviceBase interface of the concrete device classes (which is
what ZaberDeviceBase* was used for).

Make ZaberBase destructor virtual so that it actually gets called.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@13384 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd