PicardStage: Don't try to init with S.N. -1
If the CPiDetector doesn't pick up three stages, at least one stage will
be assigned a serial number of -1. Originally, the new serial number was
tested immediately in BeforeGet; this behavior blocks the option to enter
a custom serial number that the detector may have missed. (The only likely
difference in behavior is if the stage is disconnected after CPiDetector
has run, but before the Z stage has been added. Even then, it only means
the field will be filled with a no-longer-existent serial number that the
user will need to change.)

Signed-off-by: Luke Stuyvenberg <stuyvenberg@wisc.edu>

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@13743 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd