Fix up manual edit of generated code in r13296
Apply the changes to the .form, so that the changes are not lost the
next time the Java code is generated. The effect of removing the
windowClosed event from the .form is identical to the manual edit in
r13296, so there is no change to the .java in this commit.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@13797 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd