Prevent deadlock in Stage Position List
Upon stage position change notifications from the Core, the position
list was (on the same thread) calling into the Core to get the stage
positions. This can cause deadlock if some other thread is also calling
into the same stage device adapter. It was also not thread safe (curPos_
was not protected by a lock).

Fix by always processing the position change notification on the EDT.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@14160 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd