Partially revert r14500 "acqEngine: Remove rogue references to the gui's position list"
It turns out that removed functions were called from SlideExplorer2.
This kind of surprising interdependence should be eliminated, but
reverting for now.

After this commit, the acquisition engine itself continues to _not_ get
the gui's position list when passed an explicit null poistion list to
run(). Only reverting the functions that are not used within the
acquisition engine.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@14504 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd