Partially revert r14432.
It turns out that we have nasty code that has been written with
practically no regard to what thread it is running on, and explicitly
running message boxes (JOptionPane) on the EDT can cause deadlock (the
problem was seen in PixelCalibrator's Stop button, but there could be
other instances).

Existing code thus depends on ReportingUtils.showMessage() and similar
methods to block the calling thread, but not to block the EDT. This
seems to be the empirical behavior of JOptionPane (on OS X and probably
on Windows), but the documentation does not specify this (Swing is just
not supposed to be thread safe), and relying on it is dangerous. So the
ultimate fix will require that all uses of ReportingUtils be reviewed
and those calling message-box-showing methods from non-EDT threads be
fixed. Talk about technical debt.

For now, I'm reverting to relying on JOptionPane's undocumented
behavior, since it is known to "work" in the short term.

This commit also contains an update for out-of-memory error messages, so
that they continue to work after reverting the changes to the
ReportingUtils methods.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@14521 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd