[TIFF saving] Fix a bug that breaks split-by-stage-position saving.
In particular, if any image is generated with no "PositionName" tag when the
"Save XY positions in separate files" option is set, then that image will be
saved to the file for stage position 0. This can create multiple
MultipageTiffWriters for the same file, which can trample each other and
corrupt file data.

Fortunately, currently this never happens in normal operation, because separate
code ensures that the PositionName tag is always created. Still, bad practice.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@14708 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd