Avoid discarding first image in Live acquisition
We were waiting for the first image before setting up the display, but
then discarding that first image. This could result in a black image
canvas being displayed for Live with very long exposures.

This is a quick fix, and there is more to be improved as the added
comments indicate.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@14742 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd