[PI Z stage] Fix mismatch in control mode strings.
This should fix an issue where the mode string that the device adapter
reported was different from the mode string it accepted as a parameter
(e.g. when loading config files). Now it uses "Local: Frontpanel control" and
"Remote: Interface commands mode" as the displayed strings regardless of what
the device reports, and decides which to use based on if "Remote" is in the
returned string.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@14992 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd