[Multicamera colors] Fix saving of multicamera color prefs.
One of our checks prior to saving multicamera color prefs is to ensure that the
number of "data channels" matches the number of "multicamera channels".
However, this is never the case: there's always twice as many data channels
as there are multicamera channels (at least, for a two-camera multicam setup).
I simply stripped away the sanity checking as I can't see how it actually
helps.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@15124 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd