[Live] Fix excessive restarts of live mode.
When device properties are changed, we had been restarting live mode three
times; now we only restart it once. One time was to pull new device property
information from the Core, uncached; we must stop live here because some
cameras can't read properties when acquiring data. The second and third were
side-effects of refreshing the GUI. The second was "changing" the binning
mode (when the bin mode dropdown menu is repopulated); we now check that the
bin mode actually is changing before doing anything. The third was reloading
the Device Property Browser from the cache; since this is cached, it doesn't
actually touch the devices at all and we can safely do it while live mode is
running.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@15232 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd