[AcqEng] Fix stale system state cache in image tags.
Our TaggedImages were coming out of the SWIG layer with up-to-date copies of
the system state cache. However, the acquisition engine was keeping around a
stale copy of the same cache, and overwriting the images' copy with its old
copy. The engine's copy of the cache was only updated before snaps or sequence
acquisitions. Consequently, any properties that changed in the middle of a
sequence acquisition would not be reflected in the image tags. This change
removes the use of the system state cache by the acquisition engine, so that
the image's copy of the cache is preserved unmodified.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@15510 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd