LeicaDMI: Don't time out when setting to same position
For some turrents and switches, we were depending on asynchronous
(unrequested) events to go "not busy", but (at least on a DMi8) some of
those events never arrive when there is no actual mechanical movement.
So also go "not busy" when the command completion is acknowledged.

This fixes issues with the side port, magnification changer, and
objective immersion method, where "setting" to the current position
would cause the device to stay in "busy" state perpetually until the
next real position change.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@15950 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd