Generic SLM: avoid null pointer exception (Nico)
Occassional crashes were caused by an exception caused by an
uninitialized variable. Since it passed the test for NULL, it was not
properly initialized. Setting it to NULL explicitly in the constructor
avoids that situtaion (the short title of this commit is misleading).
Also added some more error testing, error reporting and error
propagation.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@16104 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd