CircularBuffer: Fix incorrectly duplicated mutexes
The two mutexes were defined as static in the header, meaning that each
compilation unit had its own copy. So inline member functions of
CircularBuffer were using a different copy of g_bufferLock when called
from MMCore.cpp vs CircularBuffer.cpp.

It is not clear if this ever actually caused symptoms, since the only
potentially critical inline function (Clear()) is usually called with
the camera lock acquired, or along with CircularBuffer::Initialize()
(which is not inline).

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@16285 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd