TCPIPPort code by Luka Lang. This device adapter can be used as serial port and forwards all communication to a server over TCP/IP. It can for example be used with a program like this one to connect to a COM port on another computer. Multiple TCPIPPorts
You can add multiple ports to a single hardware configuration. To do so, you need to use the first one for a device, then save the configuration and edit it again, This way, you get an additional port to choose from. This process would be made easier by something like #587.

I hope there are no major issues, I didn't have a lot of devices to test with.

Hoping someone finds this useful.

git-svn-id: https://valelab4.ucsf.edu/svn/micromanager2/trunk@16526 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd